added profile specific javascript, updated sharing, updated profile form to use duckForm, removed unneeded route, updated summernote version and options

+++ routes/ajax.js
@@ -1,11 +1,12 @@
-const express    = require('express');
-const pug        = require('pug');
-const isLoggedIn = require('../middleware/isLoggedIn');
-const formidable = require('../middleware/formidable');
-const pickTable  = require('../modules/pickTable');
-const s3         = require('../modules/s3');
-const sendEmail  = require('../modules/sendEmail');
-const router     = express.Router();
+const express      = require('express');
+const pug          = require('pug');
+const isLoggedIn   = require('../middleware/isLoggedIn');
+const formidable   = require('../middleware/formidable');
+const isAuthorized = require('../modules/isAuthorized');
+const pickTable    = require('../modules/pickTable');
+const s3           = require('../modules/s3');
+const sendEmail    = require('../modules/sendEmail');
+const router       = express.Router();
 
 //// database ajax calls
 	// used to see if (an) item(s) exist without returning the data of the items
@@ -24,7 +25,15 @@ const router     = express.Router();
 		res.send(items.length.toString());
 	});
 
-	router.get('/get/:table', isLoggedIn(true), (req, res) => {
+	router.get('/get/:table', (req, res) => {
+		const user = req.user;
+
+		if(!isAuthorized(user, 'beta-travel-wishlist')) {
+			console.error('User is unauthorized.');
+			res.status(401).send('User is unauthorized.');
+			return;
+		};
+
 		const field = req.query.field;
 		const value = req.query.value;
 		const findOne = req.query.findOne;
@@ -39,7 +48,15 @@ const router     = express.Router();
 		res.send(items);
 	});
 
-	router.post('/add/:table', isLoggedIn(true), (req, res) => {
+	router.post('/add/:table', (req, res) => {
+		const user = req.user;
+		
+		if(!isAuthorized(user, 'beta-travel-wishlist')) {
+			console.error('User is unauthorized.');
+			res.status(401).send('User is unauthorized.');
+			return;
+		};
+
 		const item = req.body;
 		const table = pickTable(req.params.table);
 
@@ -59,13 +76,31 @@ const router     = express.Router();
 		});
 	});
 
-	router.post('/update/:table', isLoggedIn(true), (req, res) => {
+	router.post('/update/:table', (req, res) => {
+		const user = req.user;
 		const item = req.body.item;
+		const tableName = req.params.table;
+
+		if(((tableName === 'Users') && (item.Id !== req.user.Id)) || !isAuthorized(user, 'beta-travel-wishlist')) {
+			console.error('User is unauthorized.');
+			res.status(401).send('User is unauthorized.');
+			return;
+		};
+
 		const table = pickTable(req.params.table);
 
 		if(!table){
 			console.error('Please choose a table.');
-			res.status(500).send('Please choose a table.');
+			res.status(400).send('Please choose a table.');
+			return;
+		} else if (tableName === 'Users') {
+			if (item.local && item.local.password) {
+				if (item.local.password.length > 4) {
+					item.local.password = table.generateHash(item.local.password);	
+				} else {
+					res.status(400).send({msg: 'Password must be at least 5 characters long',})
+				}
+			}
 		}
 
 		table.update(item)
@@ -79,13 +114,22 @@ const router     = express.Router();
 			});
 	});
 
-	router.post('/delete/:table', isLoggedIn(true), (req, res) => {
+	router.post('/delete/:table', (req, res) => {
+		const user = req.user;
+		
+		if(((tableName === 'Users') && (!isAuthorized(user))) || !isAuthorized(user, 'beta-travel-wishlist')) {
+			console.error('User is unauthorized.');
+			res.status(401).send('User is unauthorized.');
+			return;
+		};
+
 		const key = req.body['key'];
 		const table = pickTable(req.params.table);
 
 		if(!table){
 			console.error('Please choose a table.');
-			res.status(500).send('Please choose a table.');
+			res.status(400).send('Please choose a table.');
+			return;
 		}
 
 		table.delete(key)
@@ -99,10 +143,14 @@ const router     = express.Router();
 		});
 	});
 
-	
-
 //// s3 ajax calls
 	router.get('/getFiles', isLoggedIn(), (req, res) => {
+		if(req.query.folder.split('/')[0] !== req.user.Id && !req.user.isAdmin) {
+			res.status(401).send('Request unauthorized');
+			return;
+		}
+
+
 		s3.getFiles((files, folder, subFolders, marker, nextMarker) => {
 			res.send({files, folder, subFolders, marker, nextMarker});
 		}, req.query.folder, req.query.marker, req.query.maxKeys);
@@ -113,14 +161,16 @@ const router     = express.Router();
 	});
 
 	router.post('/upload', isLoggedIn(), formidable(), (req, res) => {
+		if(req.query.folder.split('/')[0] !== req.user.Id && !req.user.isAdmin) {
+			res.status(401).send('Request unauthorized');
+			return;
+		}
+
 		const files = req.files || [];
 		const length = files.length;
 
 		function uploadImage(file) {
-			const fileName = file.name.split('.');
-			const fileType = fileName.pop();
-
-			s3.uploadImage(file.path, {path: fileName.join('')}, function(err, versions, meta) {
+			s3.uploadImage(file.path, {path: file.name}, function(err, versions, meta) {
 				if (err) {
 					console.error(err);
 					res.status(500).send(err);
@@ -138,6 +188,11 @@ const router     = express.Router();
 	});
 
 	router.post('/deleteFiles', isLoggedIn(), (req, res) => {
+		if(req.query.folder.split('/')[0] !== req.user.Id && !req.user.isAdmin) {
+			res.status(401).send('Request unauthorized');
+			return;
+		}
+
 		const files = req.body.files;
 
 		s3.deleteFiles(files, (err) => {
